llmspell_bridge::globals

Module core

ABOUTME: Core global objects that are injected into script engines ABOUTME: Provides fundamental globals like Agent, Tool, Workflow, Hook, Event, State, Logger, Config, Security, Utils

Re-exports

pub use super::agent_global::AgentGlobal;

pub use super::tool_global::ToolGlobal;

pub use super::workflow_global::WorkflowGlobal;

Structs

ConfigGlobal

Config global object for script engines

LoggerGlobal

Logger global object for script engines

UtilsGlobal

Utils global object for script engines