llmspell_bridge::globals

Module types

ABOUTME: Core types and traits for the global object injection system ABOUTME: Defines GlobalObject trait and supporting types

Structs

GlobalContext

Shared context available to all globals during injection

GlobalMetadata

Metadata about a global object

InjectionMetrics

Performance metrics for global injection

Traits

GlobalObject

Trait for objects that can be injected as globals into script engines