llmspell_bridge::globals::types

Trait GlobalObject

pub trait GlobalObject: Send + Sync {
    // Required methods
    fn metadata(&self) -> GlobalMetadata;
    fn inject_lua(&self, lua: &Lua, context: &GlobalContext) -> Result<()>;
    fn inject_javascript(
        &self,
        ctx: &mut Context,
        context: &GlobalContext,
    ) -> Result<()>;

    // Provided methods
    fn initialize(&self, _context: &GlobalContext) -> Result<()> { ... }
    fn cleanup(&self) -> Result<()> { ... }
}

Trait for objects that can be injected as globals into script engines

Required Methods

fn metadata(&self) -> GlobalMetadata

Get metadata about this global

fn inject_lua(&self, lua: &Lua, context: &GlobalContext) -> Result<()>

Inject this global into a Lua environment

Errors

Returns an error if Lua injection fails

fn inject_javascript( &self, ctx: &mut Context, context: &GlobalContext, ) -> Result<()>

Inject this global into a JavaScript environment

Errors

Returns an error if JavaScript injection fails

Provided Methods

fn initialize(&self, _context: &GlobalContext) -> Result<()>

Initialize any resources needed by this global

Errors

Returns an error if resource initialization fails

fn cleanup(&self) -> Result<()>

Clean up any resources when the global is removed

Errors

Returns an error if cleanup fails